認証認可、画像アップロードまで最低限

+++ meguri/Controllers/WorksController.cs
@@ -5,22 +5,41 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using meguri.Data;
-using meguri.Models;
+using Meguri.Data;
+using Meguri.Models;
 using MimeKit;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authorization;
+using Meguri.Authorization;
 
-namespace meguri.Controllers {
+
+namespace Meguri.Controllers {
     public class WorksController : Controller {
         private readonly ApplicationDbContext _context;
 
-        public WorksController(ApplicationDbContext context) {
+        // 認可
+        protected IAuthorizationService AuthorizationService { get; }
+        protected UserManager<IdentityUser> UserManager { get; }
+
+        public WorksController(
+            ApplicationDbContext context,
+            IAuthorizationService authorizationService,
+            UserManager<IdentityUser> userManager
+        ) {
             _context = context;
+            UserManager = userManager;
+            AuthorizationService = authorizationService;
         }
 
         // GET: Works
         public async Task<IActionResult> Index() {
-            return View(await _context.Work.ToListAsync());
+            IList<Work> works = await _context.Work
+                .Include(w => w.Category)
+                .AsNoTracking()
+                .ToListAsync();
+
+            // return View(await _context.Work.ToListAsync());
+            return View(works);
         }
 
         // GET: Works/Details/5
@@ -49,34 +68,48 @@ namespace meguri.Controllers {
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(
-            int Id,string Name,string? Text,int ParentId,string WorkType,
-            string? Category,string? Tag1,string? Tag2,string? Tag3,
-            IFormFile? FileContent,bool Sexual,bool Violence,DateTime Created,DateTime Updated
+            int id, string name, string? text, int parentId, string workType,
+            int categoryId, string? tag1, string? tag2, string? tag3,
+            IFormFile? fileContent, bool sexual, bool violence,
+            DateTime created, DateTime updated
         ) {
             Work work = new Work();
             if (ModelState.IsValid) {
-                work.Id = Id;
-                work.Name = Name;
-                work.Text = Text;
-                work.ParentId = ParentId;
-                work.WorkType = WorkType;
-                work.Category = Category;
-                work.Tag1 = Tag1;
-                work.Tag2 = Tag2;
-                work.Tag3 = Tag3;
+                work.Id = id;
+                work.Name = name;
+                work.Text = text;
+                work.ParentId = parentId;
+                work.WorkType = workType;
+                work.CategoryId = categoryId;
+                work.Tag1 = tag1;
+                work.Tag2 = tag2;
+                work.Tag3 = tag3;
                 work.FileContent = null;
-                if (FileContent != null) {
-                    byte[] fileContent;
+                if (fileContent != null) {
+                    byte[] byteArray;
                     using (var memoryStream = new MemoryStream()) {
-                        FileContent.CopyTo(memoryStream);
-                        fileContent = memoryStream.ToArray();
+                        fileContent.CopyTo(memoryStream);
+                        byteArray = memoryStream.ToArray();
                     }
-                    work.FileContent = fileContent;
+                    work.FileContent = byteArray;
                 }
-                work.Sexual = Sexual;
+                work.Sexual = sexual;
                 work.Created = DateTime.Now;
-                work.Updated = DateTime.Now;    
-                
+                work.Updated = DateTime.Now;
+
+                var userId = UserManager.GetUserId(User);
+                var userName = UserManager.GetUserName(User);
+                work.UserId = userId;
+                work.UserName = userName;
+
+                // 認可
+                var isAuthorized = await AuthorizationService.AuthorizeAsync(
+                    User, work, WorkOperations.Create
+                );
+                if (!isAuthorized.Succeeded) {
+                    return Forbid();
+                }
+
                 _context.Add(work);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -94,6 +127,15 @@ namespace meguri.Controllers {
             if (work == null) {
                 return NotFound();
             }
+
+            // 認可
+            var isAuthorized = await AuthorizationService.AuthorizeAsync(
+                User, work, WorkOperations.Update
+            );
+            if (!isAuthorized.Succeeded) {
+                return Forbid();
+            }
+
             return View(work);
         }
 
@@ -102,25 +144,60 @@ namespace meguri.Controllers {
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Text,ParentId,WorkType,Category,Tag1,Tag2,Tag3,FileContent,Sexual,Violence,Created,Updated")] Work work) {
-            if (id != work.Id) {
+        public async Task<IActionResult> Edit(
+            int id, string name, string? text, int parentId, string workType,
+            int categoryId, string? tag1, string? tag2, string? tag3,
+            IFormFile? fileContent, bool sexual, bool violence, 
+            DateTime created, DateTime updated,
+            string userId
+        ) {
+            Work? work = _context.Work.Find(id);
+            if (work == null) {
                 return NotFound();
             }
 
-            if (ModelState.IsValid) {
-                try {
-                    _context.Update(work);
-                    await _context.SaveChangesAsync();
-                } catch (DbUpdateConcurrencyException) {
-                    if (!WorkExists(work.Id)) {
-                        return NotFound();
-                    } else {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+            // 認可
+            var isAuthorized = await AuthorizationService.AuthorizeAsync(
+                User, work, WorkOperations.Update
+            );
+            if (!isAuthorized.Succeeded) {
+                return Forbid();
             }
-            return View(work);
+
+            work.Id = id;
+            work.Name = name;
+            work.Text = text;
+            work.ParentId = parentId;
+            work.WorkType = workType;
+            work.CategoryId = categoryId;
+            work.Tag1 = tag1;
+            work.Tag2 = tag2;
+            work.Tag3 = tag3;
+            if (fileContent != null) {
+                byte[] UpdateFileContent;
+                using (var memoryStream = new MemoryStream()) {
+                    fileContent.CopyTo(memoryStream);
+                    UpdateFileContent = memoryStream.ToArray();
+                }
+                work.FileContent = UpdateFileContent;
+            }
+            work.Sexual = sexual;
+            work.Created = created;
+            work.Updated = DateTime.Now;
+            work.UserId = userId;
+
+            try {
+                _context.Update(work);
+                await _context.SaveChangesAsync();
+            } catch (DbUpdateConcurrencyException) {
+                if (!WorkExists(work.Id)) {
+                    return NotFound();
+                } else {
+                    throw;
+                }
+            }
+
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Works/Delete/5
@@ -135,6 +212,14 @@ namespace meguri.Controllers {
                 return NotFound();
             }
 
+            // 認可
+            var isAuthorized = await AuthorizationService.AuthorizeAsync(
+                User, work, WorkOperations.Delete
+            );
+            if (!isAuthorized.Succeeded) {
+                return Forbid();
+            }
+
             return View(work);
         }
 
@@ -143,6 +228,15 @@ namespace meguri.Controllers {
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id) {
             var work = await _context.Work.FindAsync(id);
+
+            // 認可
+            var isAuthorized = await AuthorizationService.AuthorizeAsync(
+                User, work, WorkOperations.Delete
+            );
+            if (!isAuthorized.Succeeded) {
+                return Forbid();
+            }
+
             if (work != null) {
                 _context.Work.Remove(work);
             }
